refactor: move testing code to own workspace
