+++ common/Cargo.toml
@@ -43,4 +43,3 @@ petgraph = { version = "0.8.1", default-features = false, features = [
 fxhash = "0.2.1"
 nalgebra = { version = "0.34.1", features = ["serde", "serde-serialize"] }
 fastrand = "2.3.0"
-approx = "0.5.1"
