+++ tests/src/common/colliders/mod.rs
@@ -1,0 +1,3 @@
+mod triangle;
+mod capsule;
+mod aabb;
